+++ src/board/viewe/viewe_panel_32480035.h
@@ -9,7 +9,7 @@
 // *INDENT-OFF*
 
 /* Set to 1 if using a custom board */
-#define ESP_PANEL_USE_CUSTOM_BOARD       (0)         // 0/1
+#define ESP_PANEL_USE_CUSTOM_BOARD       (1)         // 0/1
 
 #if ESP_PANEL_USE_CUSTOM_BOARD
 
@@ -17,7 +17,7 @@
 //////////////////////////// Please update the following macros to configure the LCD panel /////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /* Set to 1 when using an LCD panel */
-#define ESP_PANEL_USE_LCD           (0)     // 0/1
+#define ESP_PANEL_USE_LCD           (1)     // 0/1
 
 #if ESP_PANEL_USE_LCD
 /**
@@ -30,11 +30,11 @@
  *      - SPD2010
  *      - ST7262, ST7701, ST7789, ST7796, ST77916, ST77922
  */
-#define ESP_PANEL_LCD_NAME          ILI9341
+#define ESP_PANEL_LCD_NAME          ST7789
 
 /* LCD resolution in pixels */
 #define ESP_PANEL_LCD_WIDTH         (320)
-#define ESP_PANEL_LCD_HEIGHT        (240)
+#define ESP_PANEL_LCD_HEIGHT        (480)
 
 /* LCD Bus Settings */
 /**
@@ -63,15 +63,15 @@
 #if ESP_PANEL_LCD_BUS_TYPE == ESP_PANEL_BUS_TYPE_SPI
 
     #define ESP_PANEL_LCD_BUS_HOST_ID           (1)     // Typically set to 1
-    #define ESP_PANEL_LCD_SPI_IO_CS             (5)
+    #define ESP_PANEL_LCD_SPI_IO_CS             (42)
 #if !ESP_PANEL_LCD_BUS_SKIP_INIT_HOST
-    #define ESP_PANEL_LCD_SPI_IO_SCK            (7)
-    #define ESP_PANEL_LCD_SPI_IO_MOSI           (6)
+    #define ESP_PANEL_LCD_SPI_IO_SCK            (40)
+    #define ESP_PANEL_LCD_SPI_IO_MOSI           (45)
     #define ESP_PANEL_LCD_SPI_IO_MISO           (-1)    // -1 if not used
 #endif
-    #define ESP_PANEL_LCD_SPI_IO_DC             (4)
+    #define ESP_PANEL_LCD_SPI_IO_DC             (41)
     #define ESP_PANEL_LCD_SPI_MODE              (0)     // 0/1/2/3, typically set to 0
-    #define ESP_PANEL_LCD_SPI_CLK_HZ            (40 * 1000 * 1000)
+    #define ESP_PANEL_LCD_SPI_CLK_HZ            (80 * 1000 * 1000)
                                                         // Should be an integer divisor of 80M, typically set to 40M
     #define ESP_PANEL_LCD_SPI_TRANS_QUEUE_SZ    (10)    // Typically set to 10
     #define ESP_PANEL_LCD_SPI_CMD_BITS          (8)     // Typically set to 8
@@ -173,23 +173,31 @@
  *
  * There are two formats for the sequence code:
  *   1. Raw data: {command, (uint8_t []){ data0, data1, ... }, data_size, delay_ms}
- *   2. Formatter: ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(delay_ms, command, { data0, data1, ... }) and
+ *   2. Formater: ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(delay_ms, command, { data0, data1, ... }) and
  *                ESP_PANEL_LCD_CMD_WITH_NONE_PARAM(delay_ms, command)
  */
-/*
 #define ESP_PANEL_LCD_VENDOR_INIT_CMD()                                        \
     {                                                                          \
-        {0xFF, (uint8_t []){0x77, 0x01, 0x00, 0x00, 0x10}, 5, 0},              \
-        {0xC0, (uint8_t []){0x3B, 0x00}, 2, 0},                                \
-        {0xC1, (uint8_t []){0x0D, 0x02}, 2, 0},                                \
-        {0x29, (uint8_t []){0x00}, 0, 120},                                    \
-        or                                                                     \
-        ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xFF, {0x77, 0x01, 0x00, 0x00, 0x10}), \
-        ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xC0, {0x3B, 0x00}),                   \
-        ESP_PANEL_LCD_CMD_WITH_8BIT_PARAM(0, 0xC1, {0x0D, 0x02}),                   \
-        ESP_PANEL_LCD_CMD_WITH_NONE_PARAM(120, 0x29),                               \
+        {0x11, (uint8_t []){0x00}, 0, 120},           \
+        {0xF0, (uint8_t []){0xC3}, 1, 0},             \
+        {0xF0, (uint8_t []){0x96}, 1, 0},             \
+        {0x36, (uint8_t []){0x28}, 1, 0},             \
+        {0x3A, (uint8_t []){0x55}, 1, 0},             \
+        {0xB4, (uint8_t []){0x01}, 1, 0},             \
+        {0xB7, (uint8_t []){0xC6}, 1, 0},             \
+        {0xC0, (uint8_t []){0x80, 0x04}, 2, 0},       \
+        {0xC1, (uint8_t []){0x13}, 1, 0},             \
+        {0xC5, (uint8_t []){0xA7}, 1, 0},             \
+        {0xC5, (uint8_t []){0x16}, 1, 0},             \
+        {0xE8, (uint8_t []){0x40, 0x8a, 0x00, 0x00, 0x29, 0x19, 0xA5, 0x33}, 8, 0}, \
+        {0xE0, (uint8_t []){0xF0, 0x19, 0x20, 0x10, 0x11, 0x0A, 0x46, 0x44, 0x57, 0x09, 0x1A, 0x1B, 0x2A, 0x2D}, 14, 0}, \
+        {0xE1, (uint8_t []){0xF0, 0x12, 0x1A, 0x0A, 0x0C, 0x18, 0x45, 0x44, 0x56, 0x3F, 0x15, 0x11, 0x24, 0x26}, 14, 0}, \
+        {0xF0, (uint8_t []){0x3C}, 1, 0},               \
+        {0xF0, (uint8_t []){0x69}, 1, 0},               \
+        {0x21, (uint8_t []){0x00}, 0, 0},               \
+        {0x29, (uint8_t []){0x00}, 0, 50},              \
+        {0x2C, (uint8_t []){0x00}, 0, 0},               \
     }
-*/
 
 /* LCD Color Settings */
 /* LCD color depth in bits */
@@ -200,11 +208,11 @@
  *      - 1: BGR
  */
 #define ESP_PANEL_LCD_BGR_ORDER     (0)         // 0/1
-#define ESP_PANEL_LCD_INEVRT_COLOR  (0)         // 0/1
+#define ESP_PANEL_LCD_INEVRT_COLOR  (1)         // 0/1
 
 /* LCD Transformation Flags */
 #define ESP_PANEL_LCD_SWAP_XY       (0)         // 0/1
-#define ESP_PANEL_LCD_MIRROR_X      (0)         // 0/1
+#define ESP_PANEL_LCD_MIRROR_X      (1)         // 0/1
 #define ESP_PANEL_LCD_MIRROR_Y      (0)         // 0/1
 
 /* LCD Other Settings */
@@ -218,19 +226,18 @@
 //////////////////////////// Please update the following macros to configure the touch panel ///////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /* Set to 1 when using an touch panel */
-#define ESP_PANEL_USE_TOUCH         (0)         // 0/1
+#define ESP_PANEL_USE_TOUCH         (1)         // 0/1
 #if ESP_PANEL_USE_TOUCH
 /**
  * Touch controller name. Choose one of the following:
- *      - CST816S,CST820
- *      - CHSC6540
+ *      - CST816S
  *      - FT5x06
  *      - GT911, GT1151
  *      - ST1633, ST7123
  *      - TT21100
  *      - XPT2046
  */
-#define ESP_PANEL_TOUCH_NAME        TT21100
+#define ESP_PANEL_TOUCH_NAME        CST816S
 
 /* Touch resolution in pixels */
 #define ESP_PANEL_TOUCH_H_RES       (ESP_PANEL_LCD_WIDTH)   // Typically set to the same value as the width of LCD
@@ -261,8 +268,8 @@
                                                     // Typically set to 400K
     #define ESP_PANEL_TOUCH_I2C_SCL_PULLUP  (1)     // 0/1
     #define ESP_PANEL_TOUCH_I2C_SDA_PULLUP  (1)     // 0/1
-    #define ESP_PANEL_TOUCH_I2C_IO_SCL      (18)
-    #define ESP_PANEL_TOUCH_I2C_IO_SDA      (8)
+    #define ESP_PANEL_TOUCH_I2C_IO_SCL      (3)
+    #define ESP_PANEL_TOUCH_I2C_IO_SDA      (1)
 #endif
 
 #elif ESP_PANEL_TOUCH_BUS_TYPE == ESP_PANEL_BUS_TYPE_SPI
@@ -303,17 +310,17 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 ///////////////////////////// Please update the following macros to configure the backlight ////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-#define ESP_PANEL_USE_BACKLIGHT         (0)         // 0/1
+#define ESP_PANEL_USE_BACKLIGHT         (1)         // 0/1
 #if ESP_PANEL_USE_BACKLIGHT
 /* Backlight pin */
-#define ESP_PANEL_BACKLIGHT_IO          (45)        // IO num of backlight pin
+#define ESP_PANEL_BACKLIGHT_IO          (13)        // IO num of backlight pin
 #define ESP_PANEL_BACKLIGHT_ON_LEVEL    (1)         // 0: low level, 1: high level
 
 /* Set to 1 if you want to turn off the backlight after initializing the panel; otherwise, set it to turn on */
 #define ESP_PANEL_BACKLIGHT_IDLE_OFF    (0)         // 0: on, 1: off
 
 /* Set to 1 if use PWM for brightness control */
-#define ESP_PANEL_LCD_BL_USE_PWM        (1)         // 0/1
+#define ESP_PANEL_LCD_BL_USE_PWM        (0)         // 0/1
 #endif /* ESP_PANEL_USE_BACKLIGHT */
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -382,7 +389,7 @@
  */
 #define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MAJOR 0
 #define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_MINOR 2
-#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_PATCH 4
+#define ESP_PANEL_BOARD_CUSTOM_FILE_VERSION_PATCH 2
 
 #endif /* ESP_PANEL_USE_CUSTOM_BOARD */